requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
googlesearch-python==1.3.0
greenlet==3.3.1
idna==3.11
lxml==6.0.2
playwright==1.58.0
pyee==13.0.0
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
